cleanup: merge math/geo headers
