+++ libs/Math/CMakeLists.txt
@@ -8,22 +8,12 @@ add_library(
 
 		"include/math/fmt.h"
 
-			"include/math/geo/aabb.h"
-			"include/math/geo/_aabb_decl.h"
-			"include/math/geo/_aabb_impl.hpp"
-
-			"include/math/geo/circle.h"
-
-			"include/math/geo/plane.h"
-			"include/math/geo/_plane_decl.h"
-			"include/math/geo/_plane_impl.hpp"
-
-			"include/math/geo/ray.h"
-			"include/math/geo/sphere.h"
-
-			"include/math/geo/tri.h"
-			"include/math/geo/_tri_decl.h"
-			"include/math/geo/_tri_impl.hpp"
+		"include/math/geo/aabb.h"
+		"include/math/geo/circle.h"
+		"include/math/geo/plane.h"
+		"include/math/geo/ray.h"
+		"include/math/geo/sphere.h"
+		"include/math/geo/tri.h"
 
 		"include/math/literals.h"
